refactor: remove duplicate team enum, replace wrong imports
