--- src/main/java/com/codenames/codenames_backend/lobby/Team.java
@@ -1,9 +1,0 @@
-package com.codenames.codenames_backend.lobby;
-
-/**
- * Represents a team in the game.
- */
-public enum Team {
-    RED,
-    BLUE
-}
