+++ src/test/java/com/codenames/codenames_backend/lobby/controller/LobbyControllerTest.java
@@ -1,7 +1,7 @@
 package com.codenames.codenames_backend.lobby.controller;
 
 import com.codenames.codenames_backend.utility.Role;
-import com.codenames.codenames_backend.lobby.Team;
+import com.codenames.codenames_backend.utility.Team;
 import com.codenames.codenames_backend.lobby.services.LobbyService;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
